* command builders updated accourding to new proposed dotnet tool https://github.com/ElectronNET/Electron.NET/pull/162
* some parameters in build command fixed as optional accouding to https://github.com/ElectronNET/Electron.NET/blob/master/ElectronNET.CLI/Commands/BuildCommand.cs
* cake depedency market as private asset and updated to 0.32.1.
* cake.build script improvments
* unit test packages update.
* typo fixes.

+++ build.sh
@@ -6,12 +6,13 @@
 # Feel free to change this file to fit your needs.
 ##########################################################################
 
-CAKE_VERSION=0.27.2
+CAKE_VERSION=0.29.0
 
 # Define directories.
 SCRIPT_DIR=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
 TOOLS_DIR=$SCRIPT_DIR/tools
-CAKE_EXE=$TOOLS_DIR/Cake.CoreCLR.$CAKE_VERSION/cake.coreclr/$CAKE_VERSION/Cake.dll
+CAKE_ROOT= $TOOLS_DIR/cake.coreclr/
+CAKE_EXE=$TOOLS_DIR/cake.coreclr/Cake.dll
 
 # Define default arguments.
 SCRIPT="build.cake"
@@ -32,29 +33,36 @@ for i in "$@"; do
     shift
 done
 
-# Make sure the tools folder exist.
-if [ ! -d "$TOOLS_DIR" ]; then
-  mkdir "$TOOLS_DIR"
-fi
-
-# Make sure that tools.csproj exist.
-if [ ! -f "$TOOLS_DIR/tools.csproj" ]; then
-    echo "Creating tools.csproj..."
-    echo "<Project Sdk=\"Microsoft.NET.Sdk\"><PropertyGroup><OutputType>Exe</OutputType><TargetFramework>netcoreapp2.0</TargetFramework></PropertyGroup></Project>" > $TOOLS_DIR/tools.csproj
-    if [ $? -ne 0 ]; then
-        echo "An error occurred while creating tools.csproj."
-        exit 1
-    fi
-fi
-
 # Make sure that dotnet core installed.
 if ! [ -x "$(command -v dotnet)" ]; then
   echo 'Error: dotnet is not installed.' >&2
   exit 1
 fi
 
-# Add dependencies
-dotnet add $TOOLS_DIR/tools.csproj package Cake.CoreCLR -v $CAKE_VERSION --package-directory $TOOLS_DIR/Cake.CoreCLR.$CAKE_VERSION
+# Install cake if its not installed
+if [ ! -f "$CAKE_EXE" ]; then
+
+    # Make sure the tools folder exist.
+    if [ ! -d "$TOOLS_DIR" ]; then
+    mkdir "$TOOLS_DIR"
+    fi
+
+    # Make sure that tools.csproj exist.
+    if [ ! -f "$TOOLS_DIR/tools.csproj" ]; then
+        echo "Creating tools.csproj..."
+        echo "<Project Sdk=\"Microsoft.NET.Sdk\"><PropertyGroup><OutputType>Exe</OutputType><TargetFramework>netcoreapp2.0</TargetFramework></PropertyGroup></Project>" > $TOOLS_DIR/tools.csproj
+        if [ $? -ne 0 ]; then
+            echo "An error occurred while creating tools.csproj."
+            exit 1
+        fi
+    fi
+
+    # Add dependencies
+    dotnet add $TOOLS_DIR/tools.csproj package Cake.CoreCLR -v $CAKE_VERSION --package-directory $TOOLS_DIR
+    mv $TOOLS_DIR/cake.coreclr/$CAKE_VERSION/* $TOOLS_DIR/cake.coreclr/
+    rm -rf $TOOLS_DIR/cake.coreclr/$CAKE_VERSION/
+    rm -f $TOOLS_DIR/tools.csproj
+fi
 
 # Make sure that Cake has been installed.
 if [ ! -f "$CAKE_EXE" ]; then
@@ -64,7 +72,7 @@ fi
 
 # Start Cake
 if $SHOW_VERSION; then
-    dotnet $TOOLS_DIR/Cake.CoreCLR.$CAKE_VERSION/cake.coreclr/$CAKE_VERSION/Cake.dll -version
+    dotnet $TOOLS_DIR/cake.coreclr/Cake.dll -version
 else
-    dotnet $TOOLS_DIR/Cake.CoreCLR.$CAKE_VERSION/cake.coreclr/$CAKE_VERSION/Cake.dll $SCRIPT "${CAKE_ARGUMENTS[@]}"
+    dotnet $TOOLS_DIR/cake.coreclr/Cake.dll $SCRIPT "${CAKE_ARGUMENTS[@]}"
 fi
